refactor: modularize worker CLI into dedicated components for args, firebase, logging, and run management, integrate Sentry, and add a new Firebase function for check run updates.
