+++ apps/openci_worker_cli/lib/constants.dart
@@ -1,0 +1,5 @@
+const version = '0.6.9';
+
+const baseVmName = 'tahoe-base_v1.0.0';
+const sshUser = 'admin';
+const sshPassword = 'admin';
