+++ apps/openci_worker_cli/pubspec.yaml
@@ -13,8 +13,8 @@ dependencies:
   dart_firebase_admin: ^0.4.1
   googleapis: ^13.2.0
   googleapis_auth: ^1.6.0
-  http: ^1.6.0
   process_run: ^1.2.4
+  sentry: ^9.14.0
 
 dev_dependencies:
   lints: ^6.0.0
